test if timeout propagates to OkHttp
