+++ web/src/test/java/com/graphhopper/http/resources/RouteResourceClientHCTest.java
@@ -51,7 +51,6 @@ import org.junit.jupiter.params.provider.EnumSource;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.SocketTimeoutException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -140,16 +139,6 @@ public class RouteResourceClientHCTest {
         assertEquals("[0, 1]", alt.getPointsOrder().toString());
     }
 
-    @Test
-    public void testPutPOJO() {
-        ObjectNode requestJson = new ObjectMapper().createObjectNode();
-        requestJson.putPOJO("double", 1.0);
-        requestJson.putPOJO("int", 1);
-        requestJson.putPOJO("boolean", true);
-        // does not work requestJson.putPOJO("string", "test");
-        assertEquals("{\"double\":1.0,\"int\":1,\"boolean\":true}", requestJson.toString());
-    }
-
     @ParameterizedTest
     @EnumSource(value = TestParam.class)
     public void testAlternativeRoute(TestParam p) {
@@ -176,26 +165,6 @@ public class RouteResourceClientHCTest {
         isBetween(26, 31, path.getPoints().size());
         isBetween(1740, 1790, path.getDistance());
         assertTrue("Carrer Doctor Vilanova".contains(path.getDescription().get(0)), "expected: " + path.getDescription().get(0));
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = TestParam.class)
-    public void testTimeout(TestParam p) {
-        GraphHopperWeb gh = createGH(p);
-        GHRequest req = new GHRequest().
-                addPoint(new GHPoint(42.509225, 1.534728)).
-                addPoint(new GHPoint(42.512602, 1.551558)).
-                putHint("vehicle", "car");
-        GHResponse res = gh.route(req);
-        assertFalse(res.hasErrors(), "errors:" + res.getErrors().toString());
-
-        req.putHint(GraphHopperWeb.TIMEOUT, 1);
-        try {
-            gh.route(req);
-            fail();
-        } catch (RuntimeException e) {
-            assertEquals(SocketTimeoutException.class, e.getCause().getClass());
-        }
     }
 
     @ParameterizedTest
@@ -332,17 +301,6 @@ public class RouteResourceClientHCTest {
     void isBetween(double from, double to, double expected) {
         assertTrue(expected >= from, "expected value " + expected + " was smaller than limit " + from);
         assertTrue(expected <= to, "expected value " + expected + " was bigger than limit " + to);
-    }
-
-    @Test
-    public void testUnknownInstructionSign() throws IOException {
-        // Modified the sign though
-        ObjectMapper objectMapper = Jackson.newObjectMapper();
-        JsonNode json = objectMapper.readTree("{\"instructions\":[{\"distance\":1.073,\"sign\":741,\"interval\":[0,1],\"text\":\"Continue onto A 81\",\"time\":32,\"street_name\":\"A 81\"},{\"distance\":0,\"sign\":4,\"interval\":[1,1],\"text\":\"Finish!\",\"time\":0,\"street_name\":\"\"}],\"descend\":0,\"ascend\":0,\"distance\":1.073,\"bbox\":[8.676286,48.354446,8.676297,48.354453],\"weight\":0.032179,\"time\":32,\"points_encoded\":true,\"points\":\"gfcfHwq}s@}c~AAA?\",\"snapped_waypoints\":\"gfcfHwq}s@}c~AAA?\"}");
-        PathWrapper wrapper = PathWrapperDeserializer.createPathWrapper(objectMapper, json, true, true);
-
-        assertEquals(741, wrapper.getInstructions().get(0).getSign());
-        assertEquals("Continue onto A 81", wrapper.getInstructions().get(0).getName());
     }
 
     @ParameterizedTest
